+++ react/AffiliateProfileTable.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+import AffiliateProvider from './context/AffiliateProvider'
+import ProfileTable from './components/store/ProfileTable'
+
+function AffiliateProfileTable() {
+  return (
+    <AffiliateProvider>
+      <ProfileTable />
+    </AffiliateProvider>
+  )
+}
+
+export default AffiliateProfileTable